Fail early
When we receive questions for which we can't build answers, set the appropriate response code and
reply immediately without processing further questions. This is done to avoid situations where we
encounter multiple questions that we can't build answers for for different reasons but can only
respond with a single response code.